% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{mdt.feedback.no_score}
\alias{mdt.feedback.no_score}
\title{MDT feedback (no score)}
\usage{
mdt.feedback.no_score(dict = mdt::mdt_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given no feedback at the end of the test.
}
\examples{
\dontrun{
demo_mdt(feedback = mdt.feedback.no_score())}
}
