% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{mdt.feedback.simple_score}
\alias{mdt.feedback.simple_score}
\title{MDT feedback (simple score)}
\usage{
mdt.feedback.simple_score(dict = mdt::mdt_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant's score is reported at the end of the test.
}
\examples{
\dontrun{
standalone_mdt(feedback = mdt.feedback.simple_score())}
}
