% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone.R
\name{standalone_mdt}
\alias{standalone_mdt}
\title{Standalone MDT}
\usage{
standalone_mdt(
  title = "Melody discrimination test",
  admin_password = "replace-with-secure-password",
  researcher_email = NULL,
  languages = mdt_languages(),
  dict = mdt::mdt_dict,
  ...
)
}
\arguments{
\item{title}{(Scalar character) Title to display during testing.}

\item{admin_password}{(Scalar character) Password for accessing the admin panel.}

\item{researcher_email}{(Scalar character)
If not \code{NULL}, this researcher's email address is displayed
at the bottom of the screen so that online participants can ask for help.}

\item{languages}{(Character vector)
Determines the languages available to participants.
Possible languages include English (\code{"EN"}), French (\code{"FR"}), French (\code{"IT"}),
and German (informal: \code{"DE"}, formal: \code{"DE_F"}).
The first language is selected by default}

\item{dict}{The psychTestR dictionary used for internationalisation.}

\item{...}{Further arguments to be passed to \code{\link{mdt}()}.}
}
\description{
This function launches a standalone testing session for the MDT.
This can be used for data collection, either in the laboratory or online.
}
